# master-service/_input/master-service-docs/master-service-docs/openapi/master-api.openapi.yaml
openapi: 3.0.3
info:
  title: CI/CD Master Service API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /pipelines:
    get:
      summary: Список pipeline
      parameters:
        - { name: folderId, in: query, schema: { type: string, format: uuid } }
        - { name: isActive, in: query, schema: { type: boolean } }
        - { name: query, in: query, schema: { type: string } }
      responses:
        '200': { description: OK }
    post:
      summary: Создание pipeline
      responses:
        '201': { description: Created }
  /pipelines/{id}:
    get:
      summary: Pipeline со stage/job/dependency graph
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK }
    put:
      summary: Обновление pipeline
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK }
    delete:
      summary: Деактивация pipeline
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: No Content }
...
  /job-executions/{id}/logs:
    get:
      summary: Логи job execution из OpenSearch
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: tail, in: query, schema: { type: integer } }
      responses:
        '200': { description: OK }
...